improve update
